refactor: TextField에서 validState 제거, 타입 분리 및 유틸화
