+++ src/utils/validation.ts
@@ -1,0 +1,13 @@
+export type ValidState = {
+  isValid: true;
+};
+export type InvalidState = {
+  isValid: false;
+  message?: string;
+};
+
+export const createIsValidInstance: () => ValidState = () => ({ isValid: true });
+export const createIsInvalidInstance: (message?: string) => InvalidState = (message) => ({
+  isValid: false,
+  message,
+});
